requirements.txt
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
greenlet==3.0.3
gunicorn==22.0.0
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
packaging==24.1
PyYAML==6.0.1
Werkzeug==3.0.3
zipp==3.19.2
